[refactor](fe) Merge MasterOpExecutor and FEOpExecutor
